style: move method to private
